Refactor bloodlines actions and reducers
@jonnykry and I talked earlier about consolidating the reducers and
actions to make less boilerplate for the bloodlines (and dashboard)
code. This is the first attempt to remove unnecessary duplication in
the actions and reducers.

It turns out that all of the functionality of the previous code can
be easily replicated with just three action types rather than three
for each as previous. This means that we reduce code replication
by a significant amount, and consolidate our error handling to just
one or two components rather than scattering it throughout the state.

This also adds a TIMEOUT action to clear the modify state after a
specified number of milliseconds. This is useful for 'refreshing'
state between routes as well as fading error/success messages.

After this refactor, I'm under the suspicion that we could consolidate
many of our actions under these few action types and the 'modify'
reducer. The only code we'd have to write would be the action code
which makes sense since urls and parameter interpretation are similar
from request to request.