Update coinage actions
To use the new actions util and modify reducer, I switched the handling
of source post methods. Also, this reflects updates to the coinage api
which merged the create and update functionality because it doesn't really
make sense to have to distinguish. The server should handle it.

There are also some worthwhile updates to refs in PaymentSettings.js and
TriggerInput.js that (I think) are the correct way to bind refs in react
after the deprecation of string references :)