Consolidate error props
Rather than having errors nested in other store objects, this moves everything
to a single `errors` store with the error code as the key, and message as the
value.

Note: this also removes `ErrorMessage` from any component lower than Dashboard
this should help us handle the errors correctly and consistently in the future.