Update Roaster and Login Views
Previously, login and the dashboard could sometimes result in trying to send a
request before we have an auth token, this updates login to wait for a response
before changing the route.

Also, I removed all the string refs in UserSettings and roaster settings.

Finally, add a check that the user is a roaster, before going to roaster endpoints
and only show the roaster sidebar if the user is a roaster.